run.coffee: cleaned up tags and added a new configuration parameter for a default email address to be used as the Creator tag
ls.coffee: added new listeners for listing instances you created, and
instances created via chat. This is definitely work-in-progress, and
there are lots of opportunities for refactoring for reusability here

terminate.coffee: renamed it so that the command would go away. Keeping
it around, though, to use as a guide for creating a stop.coffee.
Eventually, terminate.coffee should go away entirely.