Fix 'ec2 run' so the instance name is displayed
When creating a new ec2 instance via the bot, I saw this output:

    pending    i-XXXXXXXXXX    m3.medium    10.XXX.XXX.XXX    undefined

There you can see 'undefined' where the instance name should be
displayed. My guess is that this is because the tags have not yet
been added. But we already know the name, so we can display it
without consulting the tags.