Simplify 'ec2 reserve' and 'ec2 reserve-ls' commands
The first version of the 'ec2 reserve' command was mainly intended
to be generated by code (one bot talking to another), and we did
not put much thought into its interface. It turns out that this is
not how the command is being used: people are typing it into chat,
and for that reason we wanted to make it easier to use.

The new version of 'ec2 reserve' has the following improvements:

  - detects the user issuing the command automatically; now username
    does not need to be provided

  - accepts an alias/shorthand instance ID instead of the unwieldy
    ec2 instance ID

  - no longer requires the --LongArgumentNames

  - no longer requires quotes around the comment

The 'ec2 reserve-ls' command has been updated to remove some columns
that seem in retrospect to have been superfluous. If this is wrong
they can easily be added back. It has also been updated to display
the new shorthand/alias instance IDs instead of ec2 instance IDs --
to match the corresponding change to 'ec2 reserve'.