DEV-2643 - update schedule tagging to use the new format
We now use a UTC-based schedule tag format, and are also storing
this new schedule information in a different ec2 tag ('RunSchedule'
instead of 'Schedule').

It does appear that the new format accommodates day-based scheduling,
however it's currently out of scope to add this functionality, and
we will continue to assume Monday-Friday scheduling is what's wanted,
since that's all the old instance scheduling setup supported. Future
maintainers may want to expand the scheduling functionality to allow
for more granularity -- if, say, instances need to be up over the
weekend.